# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── INVOICES ───────────────────────────────────────────────────────────────
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    patch:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    delete:
      operationId: deletePurchaseOrder
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /purchase-orders/{id}/status:
    patch:
      operationId: updatePurchaseOrderStatus
      tags: [purchase-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

  # ── DASHBOARD ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Item:
      type: object
      required: [id, name, category, unitPrice, costPrice, unitOfMeasure, createdAt]
      properties:
        id:            { type: integer }
        name:          { type: string }
        description:   { type: ["string", "null"] }
        category:      { type: string }
        unitPrice:     { type: number }
        costPrice:     { type: number }
        unitOfMeasure: { type: string }
        createdAt:     { type: string }
        updatedAt:     { type: string }
...
    Supplier:
      type: object
      required: [id, name, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        email:        { type: ["string", "null"] }
        phone:        { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        city:         { type: ["string", "null"] }
        country:      { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:          { type: string }
        email:         { type: string }
        phone:         { type: string }
        address:       { type: string }
        city:          { type: string }
        country:       { type: string }
        contactPerson: { type: string }
        notes:         { type: string }
    SupplierUpdate:
      type: object
      properties:
        name:          { type: string }
        email:         { type: string }
        phone:         { type: string }
        address:       { type: string }
        city:          { type: string }
        country:       { type: string }
        contactPerson: { type: string }
        notes:         { type: string }

    # ── INVOICE ──────────────────────────────────────────────────────────────
...
    PurchaseOrderLineItem:
      type: object
      required: [id, purchaseOrderId, itemId, description, quantity, unitCost, lineTotal]
      properties:
        id:              { type: integer }
        purchaseOrderId: { type: integer }
        itemId:          { type: ["integer", "null"] }
        description:     { type: string }
        quantity:        { type: number }
        unitCost:        { type: number }
        lineTotal:       { type: number }
        item:
          $ref: "#/components/schemas/Item"
    PurchaseOrder:
      type: object
      required: [id, poNumber, supplierId, status, subtotal, total, createdAt]
      properties:
        id:         { type: integer }
        poNumber:   { type: string }
        supplierId: { type: integer }
        status:     { type: string, enum: [Draft, Sent, Received, Cancelled] }
        orderDate:  { type: ["string", "null"] }
        expectedDate: { type: ["string", "null"] }
        subtotal:   { type: number }
        total:      { type: number }
        notes:      { type: ["string", "null"] }
        createdAt:  { type: string }
        supplier:
          $ref: "#/components/schemas/Supplier"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderLineItem"
    PurchaseOrderLineItemInput:
      type: object
      required: [description, quantity, unitCost]
      properties:
        itemId:      { type: integer }
        description: { type: string }
        quantity:    { type: number }
        unitCost:    { type: number }
    PurchaseOrderInput:
      type: object
      required: [supplierId, lineItems]
      properties:
        supplierId:   { type: integer }
        orderDate:    { type: string }
        expectedDate: { type: string }
        notes:        { type: string }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderLineItemInput"
    PurchaseOrderUpdate:
      type: object
      properties:
        supplierId:   { type: integer }
        orderDate:    { type: string }
        expectedDate: { type: string }
        notes:        { type: string }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderLineItemInput"
    PurchaseOrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [Draft, Sent, Received, Cancelled] }

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...