# lib/api-zod/src/generated/types/purchaseOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { PurchaseOrderLineItem } from './purchaseOrderLineItem';
import type { PurchaseOrderStatus } from './purchaseOrderStatus';
import type { Supplier } from './supplier';

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  supplierId: number;
  status: PurchaseOrderStatus;
  /** @nullable */
  orderDate?: string | null;
  /** @nullable */
  expectedDate?: string | null;
  subtotal: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  supplier?: Supplier;
  lineItems?: PurchaseOrderLineItem[];
}