# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
  contactPerson?: string;
  notes?: string;
}